Refactoring - moved CreatedTrelloCard.class to correct package
